refactor(web): dissolve session-socket adapters/cache into named modules
