+++ packages/web/src/lib/session-socket/event-log.test.ts
@@ -3,12 +3,10 @@ import type { SandboxEvent } from "@/types/session";
 import {
   collapseReplayTokenEvents,
   ingestLiveSandboxEvent,
-  parseWsMessage,
   pendingToTokenEvent,
-  toUiArtifact,
   toUiSandboxEvent,
   type PendingAssistantText,
-} from "./adapters";
+} from "./event-log";
 
 function tokenEvent(messageId: string, content: string, timestamp = 1): SandboxEvent {
   return { type: "token", content, messageId, sandboxId: "sb-1", timestamp };
@@ -17,21 +15,6 @@ function tokenEvent(messageId: string, content: string, timestamp = 1): SandboxE
 function completionEvent(messageId: string, timestamp = 2): SandboxEvent {
   return { type: "execution_complete", messageId, success: true, sandboxId: "sb-1", timestamp };
 }
-
-describe("parseWsMessage", () => {
-  it("returns the parsed message for a valid payload", () => {
-    expect(parseWsMessage({ type: "pong", timestamp: 123 })).toEqual({
-      type: "pong",
-      timestamp: 123,
-    });
-  });
-
-  it("returns null for unknown or malformed payloads", () => {
-    expect(parseWsMessage({ type: "not_a_message" })).toBeNull();
-    expect(parseWsMessage("garbage")).toBeNull();
-    expect(parseWsMessage(null)).toBeNull();
-  });
-});
 
 describe("toUiSandboxEvent", () => {
   it("keeps a numeric timestamp", () => {
@@ -151,83 +134,5 @@ describe("pendingToTokenEvent", () => {
     expect(
       pendingToTokenEvent({ content: "final", messageId: "msg-1", sandboxId: "sb-1", timestamp: 1 })
     ).toEqual(tokenEvent("msg-1", "final", 1));
-  });
-});
-
-describe("toUiArtifact", () => {
-  it("maps PR metadata and derives prState from tracked lifecycle over the legacy key", () => {
-    const artifact = toUiArtifact({
-      id: "artifact-pr-1",
-      type: "pr",
-      url: "https://github.com/acme/web/pull/1",
-      metadata: {
-        number: 1,
-        state: "open",
-        lifecycleState: "open",
-        isDraft: true,
-        head: "feature",
-        base: "main",
-      },
-      createdAt: 100,
-      updatedAt: 200,
-    });
-    expect(artifact).toEqual({
-      id: "artifact-pr-1",
-      type: "pr",
-      url: "https://github.com/acme/web/pull/1",
-      createdAt: 100,
-      updatedAt: 200,
-      metadata: expect.objectContaining({
-        prNumber: 1,
-        prState: "draft",
-        head: "feature",
-        base: "main",
-      }),
-    });
-  });
-
-  it("falls back to the legacy state key on artifacts without lifecycle tracking", () => {
-    const artifact = toUiArtifact({
-      id: "artifact-pr-2",
-      type: "pr",
-      url: "https://github.com/acme/web/pull/2",
-      metadata: { number: 2, state: "closed" },
-      createdAt: 100,
-    });
-    expect(artifact.metadata?.prState).toBe("closed");
-  });
-
-  it("drops wrong-type metadata fields during narrowing", () => {
-    const artifact = toUiArtifact({
-      id: "artifact-shot-1",
-      type: "screenshot",
-      url: "sessions/s/media/a.png",
-      metadata: {
-        mimeType: "image/png",
-        sizeBytes: "five",
-        viewport: "not-an-object",
-        caption: 42,
-      },
-      createdAt: 100,
-    });
-    expect(artifact.metadata).toEqual(
-      expect.objectContaining({
-        mimeType: "image/png",
-        sizeBytes: undefined,
-        viewport: undefined,
-        caption: undefined,
-      })
-    );
-  });
-
-  it("leaves metadata undefined when the artifact has none", () => {
-    const artifact = toUiArtifact({
-      id: "artifact-branch-1",
-      type: "branch",
-      url: null,
-      metadata: null,
-      createdAt: 100,
-    });
-    expect(artifact.metadata).toBeUndefined();
   });
 });
