+++ packages/web/src/lib/session-socket/swr-revalidation.ts
@@ -2,21 +2,23 @@ import { isUnarchivedSessionListKey } from "@/lib/session-list";
 import type { ServerMessage } from "@open-inspect/shared";
 
 /** An SWR cache key or key matcher to pass to `mutate`. */
-export type SessionCacheKey = string | ((key: unknown) => boolean);
+export type SwrRevalidationKey = string | ((key: unknown) => boolean);
 
 /**
- * SWR cache keys to revalidate in response to a server message. This is the
- * whole cache side-effect surface of the session socket — the hook maps each
- * key through `mutate`; everything here stays pure and testable.
+ * Which SWR caches a server message invalidates. Session-socket messages can
+ * change data that other views render — the sidebar session list and a
+ * session's child list — and this decides, per message, which of those must
+ * refetch. `useSessionSocket` maps each key through `mutate`; everything here
+ * stays pure and testable.
  *
  * Only PR artifacts revalidate the session list — they feed the sidebar's PR
  * summary; media artifacts (screenshots, video) arrive at high frequency
  * during a run and cannot change the list.
  */
-export function cacheKeysToRevalidate(
+export function swrKeysToRevalidate(
   message: ServerMessage,
   sessionId: string
-): SessionCacheKey[] {
+): SwrRevalidationKey[] {
   switch (message.type) {
     case "artifact_created":
     case "artifact_updated":
